requirements.txt
alembic==1.4.3
aniso8601==8.0.0
appdirs==1.4.3
attrs==20.3.0
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
Deprecated==1.2.10
distlib==0.3.0
distro==1.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.20
urllib3==1.25.8
vonage==2.5.4
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.2.0
